chore: add custom domain action to static site
