+++ internal/pkg/deploy/cloudformation/stack/static_site_integration_test.go
@@ -66,7 +66,8 @@ func TestStaticSiteService_TemplateAndParamsGeneration(t *testing.T) {
 			}
 			serializer, err := stack.NewStaticSite(&stack.StaticSiteConfig{
 				App: &config.Application{
-					Name: appName,
+					Name:   appName,
+					Domain: "example.com",
 				},
 				EnvManifest: envConfig,
 				Manifest:    mft.(*manifest.StaticSite),
@@ -74,6 +75,7 @@ func TestStaticSiteService_TemplateAndParamsGeneration(t *testing.T) {
 					EnvVersion: "v1.42.0",
 				},
 				AssetMappingURL: "s3://stackset-bucket/mappingfile",
+				RootUserARN:     "arn:aws:iam::123456789123:root",
 			})
 			require.NoError(t, err)
 			// validate generated template
@@ -81,7 +83,6 @@ func TestStaticSiteService_TemplateAndParamsGeneration(t *testing.T) {
 			require.NoError(t, err)
 			var actualTmpl map[any]any
 			require.NoError(t, yaml.Unmarshal([]byte(tmpl), &actualTmpl))
-
 			resetCustomResourceLocations(actualTmpl)
 
 			var expectedTmpl map[any]any
