+++ internal/pkg/deploy/cloudformation/stack/testdata/workloads/static-site.params.json
@@ -2,6 +2,7 @@
   "Parameters": {
     "AddonsTemplateURL": "",
     "AppName": "my-app",
+    "DNSDelegated": "true",
     "EnvName": "my-env",
     "WorkloadName": "static"
   },
