+++ internal/pkg/deploy/upload/customresource/customresource.go
@@ -52,7 +52,8 @@ var (
 	dnsDelegationFilePath            = path.Join(customResourcesDir, "dns-delegation.js")
 	envControllerFilePath            = path.Join(customResourcesDir, "env-controller.js")
 	nlbCertValidatorFilePath         = path.Join(customResourcesDir, "nlb-cert-validator.js")
-	nlbCustomDomainFilePath          = path.Join(customResourcesDir, "nlb-custom-domain.js")
+	nlbCustomDomainFilePath          = path.Join(customResourcesDir, "wkld-custom-domain.js")
+	staticSiteCustomDomainFilePath   = path.Join(customResourcesDir, "wkld-custom-domain.js")
 	uniqueJSONValuesFilePath         = path.Join(customResourcesDir, "unique-json-values.js")
 	triggerStateMachineFilePath      = path.Join(customResourcesDir, "trigger-state-machine.js")
 	copyAssetsFilePath               = path.Join(customResourcesDir, "copy-assets.js")
@@ -149,6 +150,7 @@ func StaticSite(fs template.Reader) ([]*CustomResource, error) {
 	return buildCustomResources(fs, map[string]string{
 		triggerStateMachineFnName: triggerStateMachineFilePath,
 		copyAssetsFnName:          copyAssetsFilePath,
+		customDomainFnName:        staticSiteCustomDomainFilePath,
 	})
 }
