+++ internal/pkg/template/templates/workloads/partials/cf/nlb.yml
@@ -124,8 +124,8 @@ NLBCustomDomainAction:
   Condition: HasAssociatedDomain
   Properties:
     ServiceToken: !GetAtt NLBCustomDomainFunction.Arn
-    LoadBalancerHostedZoneID: !GetAtt PublicNetworkLoadBalancer.CanonicalHostedZoneID
-    LoadBalancerDNS: !GetAtt PublicNetworkLoadBalancer.DNSName
+    PublicAccessHostedZoneID: !GetAtt PublicNetworkLoadBalancer.CanonicalHostedZoneID
+    PublicAccessDNS: !GetAtt PublicNetworkLoadBalancer.DNSName
     EnvHostedZoneId:
       Fn::ImportValue:
         !Sub "${AppName}-${EnvName}-HostedZone"
